requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
streamlit==1.40.1
pandas==2.2.0
numpy==1.26.4
xgboost==2.1.4
scikit-learn==1.5.0
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
joblib==1.4.2